% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcnf.R
\name{getcnf}
\alias{getcnf}
\title{Get configuration}
\usage{
getcnf(name, where = NULL)
}
\arguments{
\item{name}{(character) the name of the config}

\item{where}{(NULL|environment) where to look for the configuration}
}
\description{
Get a registered configuration.
}
\examples{
cnf::register(config = list(x = 1))
cnf::getcnf("config")
}
